Split R4 subagent tool tests
